+++ common/autoware_core_component_interface_specs/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(autoware_core_component_interface_specs)
+project(autoware_component_interface_specs)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
