+++ common/autoware_core_component_interface_specs/README.md
@@ -1,4 +1,4 @@
-# autoware_core_component_interface_specs
+# autoware_component_interface_specs
 
 This package defines the standardized component interface specifications for Autoware Core, ensuring consistent communication and interaction between various components in the Autoware autonomous driving stack.
