+++ common/autoware_core_component_interface_specs/include/autoware/core_component_interface_specs/control.hpp
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
-#define AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
+#ifndef AUTOWARE__COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
+#define AUTOWARE__COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
 
-#include <autoware/core_component_interface_specs/base.hpp>
+#include <autoware/component_interface_specs/base.hpp>
 #include <rclcpp/qos.hpp>
 
 #include <autoware_control_msgs/msg/control.hpp>
 
-namespace autoware::core_component_interface_specs::control
+namespace autoware::component_interface_specs::control
 {
 
 struct ControlCommand : InterfaceBase
@@ -32,6 +32,6 @@ struct ControlCommand : InterfaceBase
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
 };
 
-}  // namespace autoware::core_component_interface_specs::control
+}  // namespace autoware::component_interface_specs::control
 
-#endif  // AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
+#endif  // AUTOWARE__COMPONENT_INTERFACE_SPECS__CONTROL_HPP_
