+++ common/autoware_core_component_interface_specs/include/autoware/core_component_interface_specs/localization.hpp
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
-#define AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
+#ifndef AUTOWARE__COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
+#define AUTOWARE__COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
 
-#include <autoware/core_component_interface_specs/base.hpp>
+#include <autoware/component_interface_specs/base.hpp>
 #include <rclcpp/qos.hpp>
 
 #include <geometry_msgs/msg/accel_with_covariance_stamped.hpp>
 #include <nav_msgs/msg/odometry.hpp>
 
-namespace autoware::core_component_interface_specs::localization
+namespace autoware::component_interface_specs::localization
 {
 
 struct KinematicState : InterfaceBase
@@ -42,6 +42,6 @@ struct Acceleration : InterfaceBase
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_VOLATILE;
 };
 
-}  // namespace autoware::core_component_interface_specs::localization
+}  // namespace autoware::component_interface_specs::localization
 
-#endif  // AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
+#endif  // AUTOWARE__COMPONENT_INTERFACE_SPECS__LOCALIZATION_HPP_
