+++ common/autoware_core_component_interface_specs/include/autoware/core_component_interface_specs/map.hpp
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__MAP_HPP_
-#define AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__MAP_HPP_
+#ifndef AUTOWARE__COMPONENT_INTERFACE_SPECS__MAP_HPP_
+#define AUTOWARE__COMPONENT_INTERFACE_SPECS__MAP_HPP_
 
-#include <autoware/core_component_interface_specs/base.hpp>
+#include <autoware/component_interface_specs/base.hpp>
 #include <rclcpp/qos.hpp>
 
 #include <autoware_map_msgs/msg/lanelet_map_bin.hpp>
 #include <autoware_map_msgs/msg/map_projector_info.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
 
-namespace autoware::core_component_interface_specs::map
+namespace autoware::component_interface_specs::map
 {
 
 struct MapProjectorInfo : InterfaceBase
@@ -52,6 +52,6 @@ struct VectorMap : InterfaceBase
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
 };
 
-}  // namespace autoware::core_component_interface_specs::map
+}  // namespace autoware::component_interface_specs::map
 
-#endif  // AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__MAP_HPP_
+#endif  // AUTOWARE__COMPONENT_INTERFACE_SPECS__MAP_HPP_
