+++ common/autoware_core_component_interface_specs/include/autoware/core_component_interface_specs/perception.hpp
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
-#define AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
+#ifndef AUTOWARE__COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
+#define AUTOWARE__COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
 
-#include <autoware/core_component_interface_specs/base.hpp>
+#include <autoware/component_interface_specs/base.hpp>
 #include <rclcpp/qos.hpp>
 
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
 
-namespace autoware::core_component_interface_specs::perception
+namespace autoware::component_interface_specs::perception
 {
 
 struct ObjectRecognition : InterfaceBase
@@ -32,6 +32,6 @@ struct ObjectRecognition : InterfaceBase
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_VOLATILE;
 };
 
-}  // namespace autoware::core_component_interface_specs::perception
+}  // namespace autoware::component_interface_specs::perception
 
-#endif  // AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
+#endif  // AUTOWARE__COMPONENT_INTERFACE_SPECS__PERCEPTION_HPP_
