+++ common/autoware_core_component_interface_specs/include/autoware/core_component_interface_specs/vehicle.hpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
-#define AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
+#ifndef AUTOWARE__COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
+#define AUTOWARE__COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
 
 #include "base.hpp"
 
-#include <autoware/core_component_interface_specs/base.hpp>
+#include <autoware/component_interface_specs/base.hpp>
 #include <rclcpp/qos.hpp>
 
 #include <autoware_vehicle_msgs/msg/gear_report.hpp>
@@ -25,7 +25,7 @@
 #include <autoware_vehicle_msgs/msg/steering_report.hpp>
 #include <autoware_vehicle_msgs/msg/turn_indicators_report.hpp>
 
-namespace autoware::core_component_interface_specs::vehicle
+namespace autoware::component_interface_specs::vehicle
 {
 
 struct SteeringStatus : InterfaceBase
@@ -64,6 +64,6 @@ struct HazardLightStatus : InterfaceBase
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_VOLATILE;
 };
 
-}  // namespace autoware::core_component_interface_specs::vehicle
+}  // namespace autoware::component_interface_specs::vehicle
 
-#endif  // AUTOWARE__CORE_COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
+#endif  // AUTOWARE__COMPONENT_INTERFACE_SPECS__VEHICLE_HPP_
