+++ common/autoware_core_component_interface_specs/test/test_control.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/core_component_interface_specs/control.hpp"
+#include "autoware/component_interface_specs/control.hpp"
 #include "gtest/gtest.h"
 
 TEST(control, interface)
