+++ common/autoware_core_component_interface_specs/test/test_planning.cpp
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/core_component_interface_specs/planning.hpp"
+#include "autoware/component_interface_specs/planning.hpp"
 #include "gtest/gtest.h"
 
 TEST(planning, interface)
 {
   {
-    using autoware::core_component_interface_specs::planning::LaneletRoute;
+    using autoware::component_interface_specs::planning::LaneletRoute;
     LaneletRoute route;
     size_t depth = 1;
     EXPECT_EQ(route.depth, depth);
@@ -27,7 +27,7 @@ TEST(planning, interface)
   }
 
   {
-    using autoware::core_component_interface_specs::planning::Trajectory;
+    using autoware::component_interface_specs::planning::Trajectory;
     Trajectory trajectory;
     size_t depth = 1;
     EXPECT_EQ(trajectory.depth, depth);
